# API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IdleClansLeaderboardsService",
    "version": "v1"
  },
  "paths": {
...
    "/api/Clan/logs/clan/{name}": {
      "get": {
        "tags": [
          "Clan"
        ],
        "summary": "Retrieves the logs for a specific clan.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The name of the clan.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "description": "The number of logs to skip.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum number of logs to retrieve.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the requested clan logs.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Log"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Log"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Log"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If the requested limit exceeds the maximum allowed limit."
          },
          "404": {
            "description": "If no logs are found for the specified clan."
          }
        }
      }
    },
    "/api/Clan/logs/clan/{clanName}/{playerName}": {
      "get": {
        "tags": [
          "Clan"
        ],
        "summary": "Retrieves the logs for a specific player within a clan.",
        "parameters": [
          {
            "name": "clanName",
            "in": "path",
            "description": "The name of the clan.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerName",
            "in": "path",
            "description": "The name of the player.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "description": "The number of logs to skip.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum number of logs to retrieve.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the requested player logs within the clan.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Log"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Log"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Log"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If the requested limit exceeds the maximum allowed limit."
          },
          "404": {
            "description": "If no logs are found for the specified player within the clan."
          }
        }
      }
    },
...
    "/api/Clan/{clanName}/experience": {
      "get": {
        "tags": [
          "Clan"
        ],
        "summary": "Gets experience summary for a clan over a specified time period.",
        "parameters": [
          {
            "name": "clanName",
            "in": "path",
            "description": "The clan name.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "description": "Number of hours to look back (1-168, default 72).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 72
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClanExperienceSummary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClanExperienceSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClanExperienceSummary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Clan/{clanName}/experience/player/{username}": {
      "get": {
        "tags": [
          "Clan"
        ],
        "summary": "Gets a specific player's contributions within a clan.",
        "parameters": [
          {
            "name": "clanName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 72
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerClanContribution"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerClanContribution"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerClanContribution"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ClanCup/leaderboard/{gameMode}/clans/{clanName}": {
      "get": {
        "tags": [
          "ClanCup"
        ],
        "summary": "Retrieves the standing for a specific clan in a given game mode.",
        "parameters": [
          {
            "name": "gameMode",
            "in": "path",
            "description": "The game mode to retrieve the clan standing for (e.g., \"Default\", \"Ironman\").",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clanName",
            "in": "path",
            "description": "The name of the clan to retrieve the standing for.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "description": "Optional. The year to filter results by.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "description": "Optional. The month to filter results by.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Leaderboard/profile/{leaderboardName}/{name}": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Fetches the profile of an entity (player or clan) by their name.",
        "parameters": [
          {
            "name": "leaderboardName",
            "in": "path",
            "description": "Name of the leaderboard (e.g \"players:default\" to look up a player in the default game mode and \"players:ironman\" to look up an ironman).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Name of the entity.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the entity profile"
          },
          "404": {
            "description": "If the entity profile is not found"
          }
        }
      }
    },
...
    "/api/Player/profile/{name}": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Retrieves the profile for a specific player.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The name of the player.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "localize",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the requested player profile.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfile"
                }
              }
            }
          },
          "404": {
            "description": "If the player profile is not found."
          }
        }
      }
    },
...
    "/api/PlayerMarket/items/prices/latest/comprehensive/{itemId}": {
      "get": {
        "tags": [
          "PlayerMarket"
        ],
        "summary": "Retrieves detailed price information for a specific item.",
        "description": "This endpoint provides comprehensive price details for a given item, including:\r\n- The top 10 lowest prices with their corresponding volumes\r\n- The top 10 highest prices with their corresponding volumes\r\n- The average price of the item over the past 1 day, 7 days, and 30 days\r\n- The trade volume of the item over the past 1 day\r\n            \r\nThe prices and volumes are retrieved from the market offer collections and are based on the current state of the market.\r\nThe average prices and trade volume are calculated using historical trade records within the specified time ranges.\r\n            \r\nUse this endpoint to gain insights into the price distribution, average prices, and recent trade activity of an item.",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "The ID of the item to retrieve price details for.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the detailed price information for the item."
          },
          "400": {
            "description": "If the itemId is invalid or missing."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClanExperienceSummary": {
        "type": "object",
        "properties": {
          "clanName": {
            "type": "string",
            "nullable": true
          },
          "periodHours": {
            "type": "integer",
            "format": "int32"
          },
          "totalExperience": {
            "type": "number",
            "format": "double"
          },
          "skillTotals": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "playerContributions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerContribution"
            },
            "nullable": true
          },
          "intervalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Log": {
        "type": "object",
        "properties": {
          "clanName": {
            "type": "string",
            "nullable": true
          },
          "memberUsername": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "PlayerClanContribution": {
        "type": "object",
        "properties": {
          "clanName": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "periodHours": {
            "type": "integer",
            "format": "int32"
          },
          "totalExperience": {
            "type": "number",
            "format": "double"
          },
          "skillTotals": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/SkillExperience"
            },
            "nullable": true
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerExperienceInterval"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerContribution": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "totalExperience": {
            "type": "number",
            "format": "double"
          },
          "skills": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/SkillExperience"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerExperienceInterval": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "totalExperience": {
            "type": "number",
            "format": "double"
          },
          "skills": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/SkillExperience"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerProfile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "gameMode": {
            "type": "string",
            "nullable": true
          },
          "guildName": {
            "type": "string",
            "nullable": true
          },
          "skillExperiences": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "equipment": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "enchantmentBoosts": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "float"
            },
            "nullable": true
          },
          "upgrades": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "pvmStats": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "hoursOffline": {
            "type": "number",
            "format": "double"
          },
          "taskTypeOnLogout": {
            "type": "integer",
            "format": "int32"
          },
          "taskNameOnLogout": {
            "type": "string",
            "nullable": true
          },
          "activeServerId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SkillExperience": {
        "type": "object",
        "properties": {
          "experience": {
            "type": "number",
            "format": "double"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API Key needed to access protected endpoints",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  }
}